o cleanup, refactoring
