+++ nexus/nexus-core-plugins/nexus-lvo-plugin/src/main/java/org/sonatype/nexus/plugins/lvo/strategy/AbstractRemoteDiscoveryStrategy.java
@@ -22,7 +22,6 @@ import org.codehaus.plexus.component.annotations.Requirement;
 import org.codehaus.plexus.util.IOUtil;
 import org.sonatype.nexus.SystemStatus;
 import org.sonatype.nexus.configuration.application.NexusConfiguration;
-import org.sonatype.nexus.error.reporting.NexusProxyServerConfigurator;
 import org.sonatype.nexus.plugins.lvo.DiscoveryRequest;
 
 public abstract class AbstractRemoteDiscoveryStrategy
