+++ nexus/nexus-rest-api/src/main/java/org/sonatype/nexus/rest/global/GlobalConfigurationPlexusResource.java
@@ -289,7 +289,6 @@ public class GlobalConfigurationPlexusResource
                                                                     this.getActualPassword(
                                                                                             settings.getJiraPassword(),
                                                                                             getErrorReportingManager().getJIRAPassword() ) );
-                        getErrorReportingManager().setUseGlobalProxy( true );
                     }
                     else
                     {
